Pesan commit yang menjelaskan perubahan

+++ server/nlp/index.js
@@ -1,14 +1,14 @@
 const { NlpManager } = require('node-nlp');
 const Sastrawi = require('sastrawijs');
 const fs = require('fs');
-const db = require("../models");
-const path = require("path");
+const path = require('path');
 const stopwords = require('stopwords-id');
+const db = require("../models");
 
-let manager = new NlpManager({ languages: ['id'], autoSave: true, forceNER: true,  threshold: 0.3 });
+let manager = new NlpManager({ languages: ['id'], autoSave: true, forceNER: true, threshold: 0.3, useNeural: true});
 const stemmer = new Sastrawi.Stemmer();
 const modelPath = path.resolve(__basedir, "model.nlp");
-const slangwords = require('./word.json'); 
+const slangwords = require('./word.json');
 
 function replaceSlangwords(text) {
     const words = text.split(/\s+/);
@@ -27,47 +27,54 @@ function processUserInput(text) {
     const normalizedText = text.replace(/[^\w\s]|_/g, "").replace(/\s+/g, " ");
     const caseFoldedText = normalizedText.toLowerCase();
     const words = caseFoldedText.split(/\s+/);
-    const filteredWords = words.filter(word => !stopwords.includes(word));
-    const stemmedWords = filteredWords.map(word => stemmer.stem(word));
+    const stemmedWords = words.map(word => stemmer.stem(word));
     const processedText = stemmedWords.join(' ');
     return processedText;
 }
 
 async function addNerEntities() {
     const products = await db.Product.findAll({
-        attributes: ['name', 'stock', 'price'], // Tambahkan atribut 'price'
+        attributes: ['name', 'stock', 'price'],
     });
 
     products.forEach(product => {
         const productName = product.name.toLowerCase();
         const productStock = product.stock;
         const productPrice = product.price;
-        manager.addNamedEntityText(
-            'produk',
-            productName,
-            ['id'],
-            [productName]
-        );
-        manager.addNamedEntityText(
-            'stok',
-            productStock.toString(),
-            ['id'],
-            [`${productName} stok`]
-        );
-        manager.addNamedEntityText(
-            'harga',
-            productPrice.toString(),
-            ['id'],
-            [`${productName} harga`]
-        );
+
+        // Menambahkan dokumen dengan entitas stok, harga, ketersediaan, dan waktu ready
+        manager.addDocument('id', `berapa stok produk ${productName}`, 'stok');
+        manager.addDocument('id', `stok ${productName}`, 'stok');
+        manager.addDocument('id', `berapa harga produk ${productName}`, 'harga');
+        manager.addDocument('id', `harga ${productName}`, 'harga');
+        manager.addDocument('id', `apakah produk ${productName} ready`, 'ketersediaan');
+        manager.addDocument('id', `ada ${productName}`, 'ketersediaan');
+        manager.addDocument('id', `kapan ready kembali`, 'waktu_ada');
+        manager.addDocument('id', `ready ji lagi nanti`, 'waktu_ada');
+        manager.addDocument('id', `kapan ready lagi`, 'waktu_ada');
+        manager.addDocument('id', `detail produk ${productName}`, 'detail');
+        manager.addDocument('id', `lihat ${productName}`, 'detail');
+        manager.addDocument('id', `kirimkan gambar ${productName}`, 'detail');
+
+        // Menambahkan entitas produk
+        manager.addNamedEntityText('produk', productName, ['id'], [productName]);
+
+        // Menambahkan jawaban untuk setiap entitas
+        manager.addAnswer('id', 'stok', `Stok produk ${productName} adalah ${productStock}`);
+        manager.addAnswer('id', 'harga', `Harga produk ${productName} adalah ${productPrice}`);
+        manager.addAnswer('id', 'ketersediaan', productStock > 0 ? `Produk ${productName} ready kak, silahkan di order` : `Belum ready kak`);
+        manager.addAnswer('id', 'waktu_ada', `Belum tau kak, nanti kami kabari jika sudah ready kembali.`);
+        manager.addAnswer('id', 'detail', `berikut detail`);
     });
 }
 
 async function fetchDataFromDatabase() {
     try {
         const intentsData = await db.Dataset.findAll();
-        fs.unlinkSync(modelPath);
-        manager = new NlpManager({ languages: ['id'], autoSave: true, forceNER: true,  threshold: 0.3 });
+        if (fs.existsSync(modelPath)) {
+            fs.unlinkSync(modelPath);
+        }
+        manager = new NlpManager({ languages: ['id'], autoSave: true, forceNER: true, threshold: 0.3, useNeural: true });
         await addNerEntities();
         if (intentsData && intentsData.length > 0) {
             intentsData.forEach((item) => {
@@ -91,20 +98,54 @@ async function fetchDataFromDatabase() {
 async function processMessage(userMessage) {
     const processedQuestion = processUserInput(userMessage);
     const response = await manager.process('id', processedQuestion);
-    console.log('Entities:', response.entities);
-    if (response.entities && response.entities.length > 0) {
-        const productEntities = response.entities.filter(entity => entity.entity === 'produk');
-        if (productEntities.length > 0) {
-            const productName = productEntities[0].option;
+
+    // Mengambil entitas produk dari respons
+    const productEntities = response.entities.filter(entity => entity.entity === 'produk');
+
+    // Menyaring respons berdasarkan intent
+    if (response.intent === 'stok') {
+        let stockAnswers = [];
+        for (const entity of productEntities) {
+            const productName = entity.option;
             const product = await db.Product.findOne({ where: { name: productName } });
+
             if (product) {
-                response.answer += ` ${productName} ${product.stock}.`;
-                response.answer += ` ${productName} ${product.price}.`;
+                stockAnswers.push(`Stok produk ${productName} adalah ${product.stock}.`);
             } else {
-                response.answer += ` Produk ${productName} tidak ditemukan.`; 
+                stockAnswers.push(`Produk ${productName} tidak ditemukan.`);
             }
         }
+        response.answer = stockAnswers.join('\n');
+    } else if (response.intent === 'harga') {
+        let priceAnswers = [];
+        for (const entity of productEntities) {
+            const productName = entity.option;
+            const product = await db.Product.findOne({ where: { name: productName } });
+
+            if (product) {
+                priceAnswers.push(`Harga produk ${productName} adalah Rp${product.price},`);
+            } else {
+                priceAnswers.push(`Produk ${productName} tidak ditemukan.`);
+            }
+        }
+        response.answer = priceAnswers.join('\n');
+    } else if (response.intent === 'ketersediaan') {
+        let availabilityAnswers = [];
+        for (const entity of productEntities) {
+            const productName = entity.option;
+            const product = await db.Product.findOne({ where: { name: productName } });
+
+            if (product) {
+                availabilityAnswers.push(product.stock > 0 ? `Produk ${productName} ready kak, silahkan di order.` : `Produk ${productName} Belum ready kak.`);
+            } else {
+                availabilityAnswers.push(`Produk ${productName} tidak ditemukan.`);
+            }
+        }
+        response.answer = availabilityAnswers.join('\n');
+    } else if (response.intent === 'waktu_ready') {
+        response.answer = `Belum tau kak, nanti kami kabari jika sudah ready kembali.`;
     }
+
     return response;
 }
